Remove EncryptionEnumerables module and put methods in encryption class
